Update README.md
Briefly summarize The Gaming Room client and their software requirements. Who was the client? What type of software did they want you to design?
The client was The Gaming Room, they wanted their software to allow multiple users to play their game and multiple instances of the game to be run on multiple OS's.
What did you do particularly well in developing this documentation?
I documented what was expected of the software and how to achieve what was expected.
What about the process of working through a design document did you find helpful when developing the code?
I found that it was useful for organization as well as making sure that I was addressing all of the clients needs.
If you could choose one part of your work on these documents to revise, what would you pick? How would you improve it?
I would revise the recommendations section. I would put more details into why I was recommending each element.
How did you interpret the user’s needs and implement them into your software design? Why is it so important to consider the user’s needs when designing?
I took the information from the document and decided what would be most efficient for their software and made recommendations based on the most efficient and cost effective for them. It is important to do so because as software designers it is our job and responsibility to meet the customers needs.
How did you approach designing software? What techniques or strategies would you use in the future to analyze and design a similar software application?
I approached software design looking to learn how to best meet the needs of the customer. I will use the design document to breakdown the wants and needs of the customer and find the most efficient and cost effective approach to the design.